Merging r226822: ------------------------------------------------------------------------ r226822 | logan | 2015-01-22 05:38:11 -0800 (Thu, 22 Jan 2015) | 16 lines
Force unwind frame with user-defined personality.

If libcxxabi is compiled as a shared library, and the
executable references the user-defined personality routines
(e.g.  __gxx_personality_v0), then the pointer comparison in
Unwind-EHABI.cpp won't work.  This is due to the fact that
the PREL31 will point to the PLT stubs for the personality
routines (in the executable), while the __gxx_personality_v0
symbol reference is yet another (different) PLT stub (in the
libunwind.)

This will cause _Unwind_Backtrace() stops to unwind the frame
whenever it reaches __gxx_personality_v0().  This CL fix the
problem by calling the user-defined personality routines
with an undocumented API for force unwinding.

------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/libcxxabi/branches/release_36@227237 91177308-0d34-0410-b5e6-96231b3b80d8